+++ CHANGES.md
@@ -1,6 +1,7 @@
 ### Apache MetaModel (latest)
 
  * [METAMODEL-170] - Dropped support for Java 6.
+ * [METAMODEL-178] - Added AggregateFunction and ScalarFunction interfaces. Changed FunctionType enum to be super-interface of those. Compatibility is retained but a recompile of code using FunctionType is needed.
 
 ### Apache MetaModel 4.3.6